+++ packages/swirl-components/src/components/swirl-image-grid-item/swirl-image-grid-item.tsx
@@ -74,10 +74,12 @@ export class SwirlImageGridItem {
     this.gifPaused = false;
   }
 
+  componentWillLoad() {
+    this.computedSrc = this.src;
+  }
+
   componentDidLoad() {
     this.setupIntersectionObserver();
-
-    this.computedSrc = this.src;
 
     if (this.img?.complete) {
       this.loaded = true;